Extract HTMLParser to the utility
